Refacoring Storage, TrainneeDao, Training and User

+++ src/main/java/org/example/dao/TraineeDao.java
@@ -9,6 +9,6 @@ public interface TraineeDao {
     void save(Trainee trainee);
     void update(Trainee trainee);
     void deleteById(long id);
-    Optional<Trainee> findById(int id);
+    Optional<Trainee> findById(long id);
     List<Trainee> findAll();
 }
